chore(harness): phase 0 cleanup — package the app, drop dead code, add a test lane
Move the FastAPI app into the package (src/harness/ui/app.py) with
ui/server.py kept as a launcher shim so the Dockerfile CMD still works.
Delete the dead run/alk.py path (through_alk and its helpers),
voice/run_chat.py, voice/run_e2e_refactor.py, the stale design docs,
and the unused static route.

Give the harness its own test lane: bin/test-harness runs the suite in
an isolated venv, a paths-filtered harness-ci.yml lints and tests it,
and the backend pytest.ini stops collecting the directory. Adapt the
tests to the renamed mcp handler API (dual-version helpers), fix the
ruff findings the backend config surfaces, and correct the README
inaccuracies.
