+++ .github/workflows/harness-ci.yml
@@ -1,0 +1,42 @@
+name: Harness CI
+
+# The harness (futureagi/harness/) is its own package with its own test lane;
+# the backend suite deliberately does not collect it. This runs that lane.
+on:
+  pull_request:
+    paths:
+      - "futureagi/harness/**"
+      - ".github/workflows/harness-ci.yml"
+  push:
+    branches: [main, develop, feat/harness]
+    paths:
+      - "futureagi/harness/**"
+
+concurrency:
+  group: harness-ci-${{ github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    timeout-minutes: 20
+    defaults:
+      run:
+        working-directory: futureagi/harness
+    steps:
+      - uses: actions/checkout@v4
+
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+
+      - name: Install
+        run: |
+          python -m pip install --quiet --upgrade pip
+          pip install --quiet -e ".[ui,postgres,dev]"
+
+      - name: Lint
+        run: ruff check src ui tests
+
+      - name: Test
+        run: python -m pytest -q
