+++ futureagi/harness/pyproject.toml
@@ -19,14 +19,15 @@ dependencies = [
 ]
 
 [project.optional-dependencies]
-# What harness/ui/server.py needs to serve the harness over HTTP.
+# What harness.ui.app needs to serve the harness over HTTP.
 ui = [
   "fastapi>=0.115,<1",
   "uvicorn>=0.30,<1",
 ]
 # Only for a world whose store is a real Postgres rather than a file.
 postgres = ["psycopg[binary]>=3.1"]
-dev = ["pytest>=8.3", "ruff>=0.9"]
+# httpx is what fastapi.testclient drives the app with.
+dev = ["pytest>=8.3", "ruff>=0.9", "httpx>=0.27"]
 
 [tool.hatch.build.targets.wheel]
 packages = ["src/harness"]
