+++ futureagi/harness/src/harness/checks.py
@@ -23,8 +23,9 @@ two would let a typo read as a finding about the agent.
 
 from __future__ import annotations
 
+from collections.abc import Sequence
 from dataclasses import dataclass
-from typing import Any, Sequence
+from typing import Any
 
 from .world.runtime import Call, GeneratedWorld
