+++ futureagi/harness/src/harness/config.py
@@ -11,8 +11,9 @@ environment, which is also how the rest of the platform resolves them.
 from __future__ import annotations
 
 import os
+from collections.abc import Iterable
 from pathlib import Path
-from typing import Any, Iterable
+from typing import Any
 
 from claude_agent_sdk import ClaudeAgentOptions
