+++ futureagi/harness/src/harness/run/__init__.py
@@ -13,14 +13,14 @@ a hosted agent without any of this changing.
 from __future__ import annotations
 
 import json
+from collections.abc import Callable, Sequence
 from pathlib import Path
-from typing import Any, Callable, Sequence
+from typing import Any
 
-from ..contract import AgentContract
 from ..catalogue import load_catalogue
-from ..simulator import load_simulator_prompt
-from ..scenario import Scenario
+from ..contract import AgentContract
 from ..folder import apply_setup, check_ready
+from ..scenario import Scenario
 from ..world.snapshot import restore
 from .conversation import FINISHED, Exchange, Transcript, converse
 from .grade import (
@@ -34,77 +34,6 @@ from .grade import (
     summarise,
 )
 from .targets import LocalAgent, Target, register_target, resolve, supported
-
-
-def _cases(report: Any) -> list[Any]:
-    """The per-scenario results, whichever report this is.
-
-    The runner hands back a ``SimulationReport``, whose cases are ``test_cases`` and whose
-    messages live a level down; the plugins hand back the older ``TestReport``, whose cases are
-    ``results``. Reading only one of them finds nothing in the other and reports a run in which
-    nobody said anything, which is indistinguishable from an agent that ignored the person.
-    """
-    legacy = getattr(report, "to_legacy", None)
-    if callable(legacy):
-        try:
-            report = legacy()
-        except Exception:  # noqa: BLE001 - a report that will not convert is still readable
-            pass
-    return list(getattr(report, "results", None) or getattr(report, "test_cases", None) or [])
-
-
-def from_alk(report: Any, world, spent: float) -> Transcript:
-    """What ALK's report says happened, in the shape the grading already reads.
-
-    Read from ``messages``, the normalised trajectory, and not from ``transcript`` — which is a
-    string, so iterating it yields characters, matches nothing, and produces a run with no turns
-    at all. The judge is then handed an empty conversation and fails every claim about what was
-    said, which arrives looking like an agent that never spoke.
-    """
-    exchanges: list[Exchange] = []
-    for case in _cases(report):
-        for message in getattr(case, "messages", None) or []:
-            if not isinstance(message, dict):
-                continue
-            role = str(message.get("role") or "")
-            text = message.get("content") or ""
-            # Tool turns are in here too, and they are already recorded as calls. Putting them
-            # in the conversation as well would have the judge read a tool result as something
-            # the agent said.
-            if role in ("tool", "function") or not str(text).strip():
-                continue
-            exchanges.append(
-                Exchange("agent" if role == "assistant" else "customer", str(text))
-            )
-        if not exchanges and isinstance(getattr(case, "transcript", None), str):
-            # A plugin that only fills the text form still has to be readable.
-            spoken = case.transcript.strip()
-            if spoken:
-                exchanges.append(Exchange("customer", spoken))
-    return Transcript(
-        exchanges=exchanges,
-        calls=list(world.calls),
-        ended=FINISHED,
-        spent_usd=spent,
-    )
-
-
-def audio_from(report: Any) -> str:
-    """Where ALK left this run's audio, if it left any.
-
-    Artifacts are how a modality hands back what it produced, so the recording is asked of the
-    report rather than guessed at from a directory. A run with none says so.
-    """
-    for case in _cases(report):
-        for artifact in getattr(case, "artifacts", None) or []:
-            kind = str(getattr(artifact, "type", "") or "")
-            mime = str(getattr(artifact, "mime_type", "") or "")
-            if "audio" in kind.lower() or mime.startswith("audio/"):
-                found = getattr(artifact, "path", None) or getattr(artifact, "uri", None)
-                if found:
-                    return str(found)
-    return ""
-
 
 RUNS = "runs.json"
 REPORT = "report.txt"
@@ -132,7 +61,6 @@ async def run_scenario(
     *,
     target: str = "local",
     model: str | None = None,
-    through_alk: bool = False,
     on_exchange: Callable[[Exchange], Any] | None = None,
 ) -> Result:
     """Run one scenario in its own copy of the world and grade what it left behind."""
@@ -154,33 +82,15 @@ async def run_scenario(
             )
         # The setup's calls are not the agent's.
         world.calls = []
-        if through_alk:
-            # ALK owns the simulation and drives the world through EnvironmentAdapter; the
-            # harness only grades what it is left with. Nothing here is modality-specific,
-            # which is the point: the browser and voice runners take the same adapter.
-            from .alk import simulate
-
-            report, spent = await simulate(
-                scenario,
-                contract,
-                world,
-                model=model,
-                simulator_prompt=load_simulator_prompt(world_root),
-            )
-            transcript = from_alk(report, world, spent)
-            for exchange in transcript.exchanges:
-                if on_exchange:
-                    on_exchange(exchange)
-        else:
-            agent = resolve(target)(contract, world, model=model)
-            transcript = await converse(
-                agent,
-                scenario,
-                contract,
-                world_root=world_root,
-                model=model,
-                on_exchange=on_exchange,
-            )
+        agent = resolve(target)(contract, world, model=model)
+        transcript = await converse(
+            agent,
+            scenario,
+            contract,
+            world_root=world_root,
+            model=model,
+            on_exchange=on_exchange,
+        )
         # Settled by code first. The judge is only handed the sub-goals whose catalogue entry
         # says nothing observable decides them.
         settled = grade_sub_goals(world, scenario, catalogue, transcript.calls)
@@ -222,7 +132,6 @@ async def run_suite(
     *,
     target: str = "local",
     model: str | None = None,
-    through_alk: bool = False,
     out: Path | None = None,
     on_result: Callable[[Result], Any] | None = None,
     on_exchange: Callable[[Exchange], Any] | None = None,
@@ -238,7 +147,6 @@ async def run_suite(
                 world_root,
                 target=target,
                 model=model,
-                through_alk=through_alk,
                 on_exchange=on_exchange,
             )
         except Exception as failed:
