--- futureagi/harness/src/harness/run/alk.py
@@ -1,184 +1,0 @@
-"""Running a generated world through ALK's own simulation, rather than beside it.
-
-The whole reason a generated world subclasses ``EnvironmentAdapter`` is so the runners that
-already exist can drive it. ``ChatEnvironment`` takes ``environment=<adapter>`` and owns the
-synthetic user, the turn loop, the transcript and the report; the browser and voice paths take
-the same adapter. A second loop written here would work for exactly one modality and would have
-to be rewritten for the next one, which is the thing this design exists to avoid.
-
-So the split is:
-
-- **ALK** drives the simulation: who the customer is, when they speak, when it ends.
-- **The world** answers every tool call, through ``handle_tool_call``.
-- **The harness** grades afterwards, from the state the world is left in plus the transcript.
-
-What is written here is only the two adapters between the shapes: a scenario becomes an ALK
-``Persona``, and the agent under test becomes an ``AgentWrapper``.
-"""
-
-from __future__ import annotations
-
-from typing import Any
-
-from fi.simulate import Persona, Scenario as AlkScenario
-from fi.simulate.agent.wrapper import AgentInput, AgentResponse, AgentWrapper
-from fi.simulate.environments.chat import ChatEnvironment
-
-from ..contract import AgentContract
-from ..scenario import Scenario
-from ..world.runtime import GeneratedWorld
-from .targets import LocalAgent
-
-
-def as_persona(scenario: Scenario, simulator_prompt: str = "") -> Persona:
-    """One of our scenarios, in the shape ALK's simulation consumes.
-
-    ``situation`` is the simulator prompt the harness wrote for this agent with the scenario's
-    values filled in. ALK wraps it in its own voice-execution rules, so what goes here is only
-    what changes per scenario, not a second set of instructions about how to behave on a call.
-
-    The structured persona is preserved so LiveKit can vary the caller's identity, speech style,
-    and scenario metadata instead of flattening every test into the same generic customer.
-    """
-    from ..simulator import fill
-
-    # The LiveKit voice simulator treats ``situation`` as private context rather than a
-    # line to recite. Preserve the harness-authored caller rules here: they explicitly keep
-    # the simulator in the customer role and stop it from volunteering held-back details.
-    # Fall back to the plain scenario instruction for older sessions without a prompt.
-    filled = fill(simulator_prompt, scenario.slots())[0] if simulator_prompt else ""
-    persona = (
-        scenario.persona.model_dump(exclude_none=True)
-        if scenario.persona is not None
-        else {"name": "customer"}
-    )
-    return Persona(
-        persona=persona,
-        situation=filled or scenario.instruction,
-        # Said in the person's own terms, because it is read aloud with the situation.
-        # ``scenario.tests`` describes what the suite is checking — "agent correctly counts
-        # customers filtered by country" — and a person who opens by announcing what the agent
-        # is being graded on has told it the answer.
-        outcome="get what you came for, or accept that you cannot",
-    )
-
-
-def as_alk_scenario(
-    scenarios: list[Scenario], name: str = "harness", simulator_prompt: str = ""
-) -> AlkScenario:
-    return AlkScenario(
-        name=name,
-        description="generated by the harness",
-        dataset=[as_persona(one, simulator_prompt) for one in scenarios],
-    )
-
-
-def _spoken(input: AgentInput) -> str:
-    """What the customer just said, as text.
-
-    ALK passes a message as a mapping, not a string, and hands the whole history alongside it.
-    Passing the mapping straight to a session that expects text fails inside the SDK with a
-    redacted TypeError, which says nothing about where it came from.
-    """
-    latest = input.new_message
-    if isinstance(latest, dict):
-        content = latest.get("content")
-        if isinstance(content, list):
-            content = " ".join(
-                part.get("text", "") for part in content if isinstance(part, dict)
-            )
-        if content:
-            return str(content)
-    if isinstance(latest, str) and latest:
-        return latest
-    for message in reversed(input.messages or []):
-        if isinstance(message, dict) and message.get("role") != "assistant":
-            content = message.get("content")
-            if content:
-                return str(content)
-    return "(the customer said nothing)"
-
-
-class ContractAgent(AgentWrapper):
-    """The agent under test, in the shape ALK drives agents by.
-
-    It holds the same session ``LocalAgent`` uses, so the agent being graded is identical either
-    way; what changes is who runs the conversation around it. The tool calls it made are reported
-    back to ALK so they appear in the transcript, having already gone through the world.
-    """
-
-    def __init__(
-        self,
-        contract: AgentContract,
-        world: GeneratedWorld,
-        *,
-        model: str | None = None,
-    ) -> None:
-        self.agent = LocalAgent(contract, world, model=model)
-        self.world = world
-        self._open = False
-
-    async def call(self, input: AgentInput) -> AgentResponse:
-        if not self._open:
-            await self.agent.open()
-            self._open = True
-
-        before = len(self.world.calls)
-        said = await self.agent.say(_spoken(input))
-        made = self.world.calls[before:]
-
-        return AgentResponse(
-            content=said,
-            tool_calls=[
-                {"name": call.name, "arguments": call.arguments} for call in made
-            ],
-            tool_responses=[
-                {
-                    "name": call.name,
-                    "content": call.error if not call.ok else str(call.result),
-                    "success": call.ok,
-                }
-                for call in made
-            ],
-        )
-
-    async def aclose(self) -> None:
-        if self._open:
-            await self.agent.close()
-            self._open = False
-
-    @property
-    def spent_usd(self) -> float:
-        return self.agent.spent_usd
-
-
-async def simulate(
-    scenario: Scenario,
-    contract: AgentContract,
-    world: GeneratedWorld,
-    *,
-    model: str | None = None,
-    simulator_prompt: str = "",
-) -> tuple[Any, float]:
-    """Run one scenario through ALK's chat simulation against this world.
-
-    ``auto_execute_tools`` is off because the agent's tools are bound to the world already and
-    have run by the time it answers. Turning it on would execute every call a second time, which
-    for a world that really writes rows means every order placed twice.
-    """
-    agent = ContractAgent(contract, world, model=model)
-    try:
-        report = await ChatEnvironment().run(
-            scenario=as_alk_scenario(
-                [scenario], name=scenario.name, simulator_prompt=simulator_prompt
-            ),
-            agent_callback=agent,
-            environment=world,
-            auto_execute_tools=False,
-            max_turns=max(2, scenario.max_turns),
-            min_turns=2,
-            modality=contract.modality or "text",
-        )
-    finally:
-        await agent.aclose()
-    return report, agent.spent_usd
