+++ futureagi/harness/src/harness/run/conversation.py
@@ -11,8 +11,9 @@ All three are recorded, because how a conversation ended is often the finding.
 
 from __future__ import annotations
 
+from collections.abc import Callable
 from dataclasses import dataclass, field
-from typing import Any, Callable
+from typing import Any
 
 from claude_agent_sdk import ClaudeAgentOptions
