+++ futureagi/harness/src/harness/world/expectations.py
@@ -12,7 +12,8 @@ whether a particular value is among them.
 
 from __future__ import annotations
 
-from typing import Any, Mapping
+from collections.abc import Mapping
+from typing import Any
 
 COUNT = "count"
