+++ futureagi/harness/src/harness/world/probe.py
@@ -20,8 +20,9 @@ outcomes here.
 from __future__ import annotations
 
 import re
+from collections.abc import Iterable, Mapping, Sequence
 from dataclasses import dataclass, field
-from typing import Any, Iterable, Mapping, Sequence
+from typing import Any
 
 from ..contract import AgentContract, ToolSpec
 from .expectations import check_state
