+++ futureagi/harness/src/harness/world/snapshot.py
@@ -14,8 +14,9 @@ from __future__ import annotations
 import json
 import shutil
 import sqlite3
+from collections.abc import Mapping
 from pathlib import Path
-from typing import Any, Mapping
+from typing import Any
 
 from .runtime import GeneratedWorld
