+++ futureagi/harness/src/harness/world/stores/inprocess.py
@@ -19,8 +19,9 @@ from __future__ import annotations
 
 import copy
 import json
+from collections.abc import Callable, Mapping, Sequence
 from pathlib import Path
-from typing import Any, Callable, Mapping, Sequence
+from typing import Any
 
 from . import Snapshot, StoreError
