+++ futureagi/pytest.ini
@@ -18,6 +18,7 @@ norecursedirs =
     scripts
     __pycache__
     *.egg-info
+    harness
 
 python_files = test_*.py
 python_classes = Test*
